M2.5: time.Time, strings.Builder/bytes.Buffer, io.WriteString, fmt.Fprint*, shim vars
stdlib overlay batch:
- time.Time as an opaque UTC value type: Now/Unix/Date/Since construction;
  Year..Nanosecond/Weekday/Unix* accessors; Add/Sub/Before/After/Equal/IsZero;
  Format with the Go reference layout (RFC3339, fractions, AM/PM, tz). Duration
  methods unchanged.
- strings.Builder and bytes.Buffer (Write*/String/Len/Reset/Grow) via the opaque
  value-type pattern; io.WriteString; fmt.Fprint/Fprintln/Fprintf writing to a
  builder/buffer/stdout/stderr through Fmt.WriteTo.
- &opaque resolves to the same runtime handle (goType pointer-to-opaque, addrOf,
  analyzeAddrTaken skip), so &buf works without a cell.
- shim variables: os.Stdout/os.Stderr (GoFile) and time.UTC/time.Local via
  accessor externs (shimVarRegistry/shimVarExtern).

Fixtures 282_buffer, 283_time. 82 conformance fixtures, all byte-exact.
Docs: ROADMAP, ROADMAP-M2.5, GAPS updated with current P0 status.

+++ runtime/stdlib/Bytes.cs
@@ -43,7 +43,7 @@ public static class Bytes
     public static bool HasPrefix(GoSlice s, GoSlice p) => B(s).AsSpan().StartsWith(B(p));
     public static bool HasSuffix(GoSlice s, GoSlice p) => B(s).AsSpan().EndsWith(B(p));
     public static long Index(GoSlice s, GoSlice sub) => Idx(B(s), B(sub));
-    public static long IndexByte(GoSlice s, long c)
+    public static long IndexByte(GoSlice s, int c)
     {
         var b = B(s);
         for (int i = 0; i < b.Length; i++) if (b[i] == (byte)c) return i;
